requirements.txt
# Core dependencies
langchain>=0.0.350
langgraph>=0.0.60
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Data processing and storage
pydantic>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# Evaluation metrics
rouge-score>=0.1.2
bert-score>=0.3.13
nltk>=3.8.1

# Utilities
python-json-logger>=2.0.0
typing-extensions>=4.8.0
asyncio>=3.4.3

# Testing and development
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0